python-backend (қосымша арты)/requirements.txt
fastapi
uvicorn[standard]
pydantic
sqlalchemy
python-multipart
psycopg2-binary
uvicorn[standard] 
pydantic-settings
passlib[bcrypt] 
python-jose[cryptography] 
SQLAlchemy
